+++ internal/peer/common/peerclient.go
@@ -75,11 +75,11 @@ func NewPeerClientForAddress(address, tlsRootCertFile string) (*PeerClient, erro
 
 	clientConfig.MaxRecvMsgSize = comm.DefaultMaxRecvMsgSize
 	if viper.IsSet("peer.maxRecvMsgSize") {
-		clientConfig.MaxRecvMsgSize = int(viper.GetInt32("peer.maxRecvMsgSize"))
+		clientConfig.MaxRecvMsgSize = viper.GetInt64("peer.maxRecvMsgSize")
 	}
 	clientConfig.MaxSendMsgSize = comm.DefaultMaxSendMsgSize
 	if viper.IsSet("peer.maxSendMsgSize") {
-		clientConfig.MaxSendMsgSize = int(viper.GetInt32("peer.maxSendMsgSize"))
+		clientConfig.MaxSendMsgSize = viper.GetInt64("peer.maxSendMsgSize")
 	}
 
 	return newPeerClientForClientConfig(address, clientConfig)
